+++ src/aks-preview/HISTORY.rst
@@ -11,6 +11,7 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+* `az aks get-credentials``: Convert device code mode kubeconfig to Azure CLI token format to bypass conditional access login blocks.
 
 19.0.0b4
 +++++++